feat: GPT 답변 기능 추가 (#65)

+++ inpeak/src/main/java/com/blooming/inpeak/answer/dto/response/AnswerResponse.java
@@ -9,7 +9,7 @@ import lombok.Builder;
 public record AnswerResponse(
     LocalDate dateTime,
     String questionContent,
-    Long runningTime,
+    int runningTime,
     AnswerStatus answerStatus,
     boolean isUnderstood
 ) {
